update_category

+++ CameraServicesPlatform/CameraServicesPlatform.BackEnd.Application/Service/CategoryService.cs
@@ -6,6 +6,7 @@ using CameraServicesPlatform.BackEnd.Domain.Models;
 using Microsoft.AspNetCore.Identity;
 using System;
 using System.Collections.Generic;
+using System.Drawing.Printing;
 using System.Linq;
 using System.Linq.Expressions;
 using System.Text;
@@ -16,17 +17,21 @@ namespace CameraServicesPlatform.BackEnd.Application.Service
     public class CategoryService : GenericBackendService, ICategoryService
     {
         private readonly IRepository<Category> _repository;
+        private readonly IRepository<Product> _repositoryProduct;
+
         private readonly IMapper _mapper;
         private readonly IUnitOfWork _unitOfWork;
         
         public CategoryService(
             IRepository<Category> categoryRepository,
+            IRepository<Product> repositoryProduct,
             IUnitOfWork unitOfWork,
             IMapper mapper,
             IServiceProvider serviceProvider
         ) : base(serviceProvider)
         {
             _repository = categoryRepository;
+            _repositoryProduct = repositoryProduct;
             _unitOfWork = unitOfWork;
             _mapper = mapper;
         }
@@ -42,8 +47,8 @@ namespace CameraServicesPlatform.BackEnd.Application.Service
                 {
                     CategoryID = Guid.NewGuid(),
                     CategoryName = categoryResponse.CategoryName,
-                    CategoryDescription = categoryResponse.CategoryDescription
-                    
+                    CategoryDescription = categoryResponse.CategoryDescription,
+                    StatusCategory = true
                 };
 
                 await listCategory.Insert(category);
@@ -68,7 +73,8 @@ namespace CameraServicesPlatform.BackEnd.Application.Service
 
             try
             {
-                Expression<Func<Category, bool>>? filter = null;
+                Expression<Func<Category, bool>>? filter = a => a.StatusCategory == true;
+                List<CategoryDto> listCategory = new List<CategoryDto>();
 
                 var pagedResult = await _repository.GetAllDataByExpression(
                     filter,
@@ -78,9 +84,18 @@ namespace CameraServicesPlatform.BackEnd.Application.Service
                     isAscending: true,
                     null
                 );
+                foreach (var item in pagedResult.Items)
+                {
 
-                // Set success result
-                result.Result = pagedResult;
+                    CategoryDto newCategory = new CategoryDto
+                    {
+                        CategoryID = item.CategoryID.ToString(),
+                        CategoryName = item.CategoryName,
+                        CategoryDescription = item.CategoryDescription,
+                    };
+                    listCategory.Add(newCategory);
+                }
+                result.Result = listCategory;
                 result.IsSuccess = true;
             }
             catch (Exception ex)
@@ -106,15 +121,20 @@ namespace CameraServicesPlatform.BackEnd.Application.Service
                 }
 
                 var pagedResult = await _repository.GetAllDataByExpression(
-                    a => a.CategoryID == categoryId,
+                    a => a.CategoryID == categoryId && a.StatusCategory == true,
                     pageIndex,
                     pageSize,
                     orderBy: a => a.CategoryName,
                     isAscending: true,
                     null
                 );
-
-                result.Result = pagedResult;
+                CategoryDto newCategory = new CategoryDto
+                {
+                    CategoryID = pagedResult.Items[0].CategoryID.ToString(),
+                    CategoryName = pagedResult.Items[0].CategoryName,
+                    CategoryDescription = pagedResult.Items[0].CategoryDescription,
+                };
+                result.Result = newCategory;
                 result.IsSuccess = true;
             }
             catch (Exception ex)
@@ -134,8 +154,9 @@ namespace CameraServicesPlatform.BackEnd.Application.Service
 
                 if (!string.IsNullOrEmpty(categoryNameFilter))
                 {
-                    filter = a => a.CategoryName.Contains(categoryNameFilter);
+                    filter = a => a.CategoryName.Contains(categoryNameFilter) && a.StatusCategory == true;
                 }
+                List<CategoryDto> listCategory = new List<CategoryDto>();
 
                 var pagedResult = await _repository.GetAllDataByExpression(
                     filter,
@@ -145,8 +166,18 @@ namespace CameraServicesPlatform.BackEnd.Application.Service
                     isAscending: true,
                     null
                 );
+                foreach (var item in pagedResult.Items)
+                {
 
-                result.Result = pagedResult;
+                    CategoryDto newCategory = new CategoryDto
+                    {
+                        CategoryID = item.CategoryID.ToString(),
+                        CategoryName = item.CategoryName,
+                        CategoryDescription = item.CategoryDescription,
+                    };
+                    listCategory.Add(newCategory);
+                }
+                result.Result = listCategory;
                 result.IsSuccess = true;
             }
             catch (Exception ex)
@@ -175,11 +206,18 @@ namespace CameraServicesPlatform.BackEnd.Application.Service
                 categoryExist.CategoryName = categoryResponse.CategoryName;
                 categoryExist.CategoryDescription = categoryResponse.CategoryDescription;
 
+                CategoryDto categoryUpdate = new CategoryDto
+                {
+                    CategoryID = categoryExist.CategoryID.ToString(),
+                    CategoryName = categoryExist.CategoryName,
+                    CategoryDescription = categoryExist.CategoryDescription,
+                };
+
                 await categoryRepository.Update(categoryExist);
 
                 await _unitOfWork.SaveChangesAsync();
 
-                result.Result = categoryExist;
+                result.Result = categoryUpdate;
                 result.IsSuccess = true;
             }
             catch (Exception ex)
@@ -197,6 +235,23 @@ namespace CameraServicesPlatform.BackEnd.Application.Service
             {
                 var categoryRepository = Resolve<IRepository<Category>>();
                 Guid.TryParse(categoryId, out Guid id);
+                Category categoryExist = await categoryRepository.GetById(id);
+                var productExist = await _repositoryProduct.GetAllDataByExpression(
+                    a => a.CategoryID == id,
+                    1,
+                    10,
+                    null,
+                    isAscending: true,
+                    null
+                );
+                if(productExist.Items.Count() > 0)
+                {
+                    categoryExist.StatusCategory = false;
+                    await categoryRepository.Update(categoryExist);
+                    await _unitOfWork.SaveChangesAsync();
+                    result.Result = "Category deleted successfully.";
+                    return result;
+                }
                 await categoryRepository.DeleteById(id);
                 await _unitOfWork.SaveChangesAsync();
                 result.IsSuccess = true;
